+++ python/lsst/ts/mtdome/mock_llc/lcs.py
@@ -28,6 +28,7 @@ from typing import List
 from lsst.ts import utils
 from .base_mock_llc import BaseMockStatus
 from ..enums import LlcMotionState
+from lsst.ts.idl.enums.MTDome import OperationalMode
 
 NUM_LOUVERS = 34
 _NUM_MOTORS = 68
@@ -44,9 +45,11 @@ class LcsStatus(BaseMockStatus):
     def __init__(self) -> None:
         super().__init__()
         self.log = logging.getLogger("MockLcsStatus")
-        # variables holding the status of the mock Louvres
+
+        # Variables holding the status of the mock Louvres
         self.status = np.full(NUM_LOUVERS, LlcMotionState.CLOSED.name, dtype=object)
-        self.error = [{"code": 0, "description": "No Errors"}]
+        self.operational_mode = OperationalMode.NORMAL
+        self.messages = [{"code": 0, "description": "No Errors"}]
         self.position_actual = np.zeros(NUM_LOUVERS, dtype=float)
         self.position_commanded = np.zeros(NUM_LOUVERS, dtype=float)
         self.drive_torque_actual = np.zeros(_NUM_MOTORS, dtype=float)
@@ -67,7 +70,11 @@ class LcsStatus(BaseMockStatus):
             f"time_diff = {time_diff}"
         )
         self.llc_status = {
-            "status": {"error": self.error, "status": self.status.tolist()},
+            "status": {
+                "messages": self.messages,
+                "status": self.status.tolist(),
+                "operationalMode": self.operational_mode.name,
+            },
             "positionActual": self.position_actual.tolist(),
             "positionCommanded": self.position_commanded.tolist(),
             "driveTorqueActual": self.drive_torque_actual.tolist(),
@@ -118,6 +125,24 @@ class LcsStatus(BaseMockStatus):
         """Stop louvers motion and engage the brakes."""
         self.command_time_tai = utils.current_tai()
         self.status[:] = LlcMotionState.STATIONARY.name
+
+    async def set_normal(self) -> None:
+        """Set louvers motion state to normal (as opposed to
+        degraded).
+        """
+        if self.operational_mode == OperationalMode.DEGRADED:
+            self.operational_mode = OperationalMode.NORMAL
+        else:
+            self.log.warning("Operational state already is NORMAL. Ignoring.")
+
+    async def set_degraded(self) -> None:
+        """Set louvers motion state to degraded (as opposed to
+        normal).
+        """
+        if self.operational_mode == OperationalMode.NORMAL:
+            self.operational_mode = OperationalMode.DEGRADED
+        else:
+            self.log.warning("Operational state already is DEGRADED. Ignoring.")
 
     async def exit_fault(self) -> None:
         """Clear the fault state."""