+++ python/lsst/ts/mtdome/mock_llc/thcs.py
@@ -27,6 +27,7 @@ import numpy as np
 from lsst.ts import utils
 from .base_mock_llc import BaseMockStatus
 from ..enums import LlcMotionState
+from lsst.ts.idl.enums.MTDome import OperationalMode
 
 NUM_THERMO_SENSORS = 13
 
@@ -40,7 +41,8 @@ class ThcsStatus(BaseMockStatus):
         super().__init__()
         self.log = logging.getLogger("MockThcsStatus")
         self.status = LlcMotionState.CLOSED
-        self.error = [{"code": 0, "description": "No Errors"}]
+        self.messages = [{"code": 0, "description": "No Errors"}]
+        self.operational_mode = OperationalMode.NORMAL
         self.temperature = np.zeros(NUM_THERMO_SENSORS, dtype=float)
 
     async def determine_status(self, current_tai: float) -> None:
@@ -53,7 +55,11 @@ class ThcsStatus(BaseMockStatus):
             f"time_diff = {time_diff}"
         )
         self.llc_status = {
-            "status": {"error": self.error, "status": self.status.name},
+            "status": {
+                "messages": self.messages,
+                "status": self.status.name,
+                "operationalMode": self.operational_mode.name,
+            },
             "temperature": self.temperature.tolist(),
             "timestampUTC": current_tai,
         }