+++ python/lsst/ts/mtdome/schema/command.py
@@ -55,6 +55,14 @@ registry["command"] = json.loads(
         "goStationaryEl",
         "goStationaryLouvers",
         "goStationaryShutter",
+        "setNormalAz",
+        "setNormalEl",
+        "setNormalLouvers",
+        "setNormalShutter",
+        "setDegradedAz",
+        "setDegradedEl",
+        "setDegradedLouvers",
+        "setDegradedShutter",
         "setTemperature",
         "fans",
         "inflate",
@@ -505,6 +513,142 @@ registry["command"] = json.loads(
       "if": {
         "properties": {
           "command": {
+            "const": "setNormalAz"
+          }
+        }
+      },
+      "then": {
+        "properties": {
+          "parameters": {
+            "type": "object",
+            "additionalProperties": false
+          }
+        }
+      }
+    },
+    {
+      "if": {
+        "properties": {
+          "command": {
+            "const": "setNormalEl"
+          }
+        }
+      },
+      "then": {
+        "properties": {
+          "parameters": {
+            "type": "object",
+            "additionalProperties": false
+          }
+        }
+      }
+    },
+    {
+      "if": {
+        "properties": {
+          "command": {
+            "const": "setNormalLouvers"
+          }
+        }
+      },
+      "then": {
+        "properties": {
+          "parameters": {
+            "type": "object",
+            "additionalProperties": false
+          }
+        }
+      }
+    },
+    {
+      "if": {
+        "properties": {
+          "command": {
+            "const": "setNormalShutter"
+          }
+        }
+      },
+      "then": {
+        "properties": {
+          "parameters": {
+            "type": "object",
+            "additionalProperties": false
+          }
+        }
+      }
+    },
+    {
+      "if": {
+        "properties": {
+          "command": {
+            "const": "setDegradedAz"
+          }
+        }
+      },
+      "then": {
+        "properties": {
+          "parameters": {
+            "type": "object",
+            "additionalProperties": false
+          }
+        }
+      }
+    },
+    {
+      "if": {
+        "properties": {
+          "command": {
+            "const": "setDegradedEl"
+          }
+        }
+      },
+      "then": {
+        "properties": {
+          "parameters": {
+            "type": "object",
+            "additionalProperties": false
+          }
+        }
+      }
+    },
+    {
+      "if": {
+        "properties": {
+          "command": {
+            "const": "setDegradedLouvers"
+          }
+        }
+      },
+      "then": {
+        "properties": {
+          "parameters": {
+            "type": "object",
+            "additionalProperties": false
+          }
+        }
+      }
+    },
+    {
+      "if": {
+        "properties": {
+          "command": {
+            "const": "setDegradedShutter"
+          }
+        }
+      },
+      "then": {
+        "properties": {
+          "parameters": {
+            "type": "object",
+            "additionalProperties": false
+          }
+        }
+      }
+    },
+    {
+      "if": {
+        "properties": {
+          "command": {
             "const": "setTemperature"
           }
         }
@@ -696,6 +840,5 @@ registry["command"] = json.loads(
       }
     }
   ]
-}
-    """
+}    """
 )