+++ doc/version_history.rst
@@ -14,6 +14,9 @@ Changes:
 * Modified the unit tests because of changes in IDL.
 * Replaced the use of ts_salobj functions with ts_utils functions.
 * Added auto-enable capability.
+* Renamed "error" to "messages" in the status telemetry.
+* Added "operationalMode" to the status telemetry.
+* Added commands to change the operational mode of a lower level component.
 
 Requires: